Configure cookie path for q_session cookie.
While Gecko defaults to / as the cookie path if not set, Safari
defaults to the page that set the cookie.  This means that the session
cookie is only valid for the page a login flow was started from but
not any other pages and so a new login is required for each page,
which makes no sense and is very inconvenient.  This patch fixes the
problem by explicity setting the cookie path to /.

Change-Id: I8edc59958fe9120f70af1c00ebc8ae103ef756f7