KB56 Fix resource copy step.
The Maven Resources Plugin defaulted to not overwriting files that
were already present in the target if they had a newer modification
timestamp.  This led to unpredictable behavior during the packaging
phase, where the resources generated by Parcel did not override the
original source files.

This patch sets the overwrite flag when copying the generated
resources, fixing the problem.

Change-Id: I047f5eb7d56e4444233ec12d94be43a12137010b